fix: tighten dev startup cleanup
