+++ .changeset/dev-startup-optimizations.md
@@ -2,4 +2,4 @@
 'rsbuild-plugin-react-router': patch
 ---
 
-Improve development startup performance with guarded lazy compilation defaults, configurable route transform workers, optional lazy-compilation prewarming, and leaner dev build orchestration.
+Improve development startup performance with guarded lazy compilation enabled by default, configurable route transform workers, unstable opt-in lazy-compilation prewarming, and leaner dev build orchestration.
